+++ changelog.md
@@ -3,6 +3,7 @@
 # v1.2.0 (Unreleased)
 ### New Features
 * (#9) Revision list entries now have a context menu that allows more ways of interacting with the revision.
+* (#48) Added Settings window.
 ### Bugfix
 * (#42) Fixed issue that prevented correct scans when using a directory referenced by UNC-paths.
 * (#45) Commit time in revision head area now shows local time instead of UTC.